Fix grammar and completion loading problems for all
- As nyteshade mentioned in #1 sometimes grammar files are loaded twice
(if not thrice) because of racing conditions (when the package is
installed firstly or after a change in highlight preferences). This
commit fixes it completely by renaming the directory "grammars" as
"grammar". For more information refer to the readme in the "grammar"
directory.
- Similarly the fix for #2 contained a bug that causes the completions
loaded twice after a change in highlight preferences. This commit fixes
it completely.
- This commit implements precautionally a temporary function to delete
any possible leftover "grammars" directories created by previous
versions (if not removed by package updater). A note to myself! DON'T
FORGET TO REMOVE IT IN THE FUTURE!